requirements.txt
async-timeout==4.0.2
black==22.6.0
click==8.1.3
Deprecated==1.2.13
deta==1.1.0
gitdb==4.0.9
GitPython==3.1.27
humanize==4.2.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pyparsing==3.0.9
redis==4.3.4
smmap==5.0.0
tomli==2.0.1
wrapt==1.14.1
